% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialstore-package.R
\docType{package}
\name{spatialstore-package}
\alias{spatialstore}
\alias{spatialstore-package}
\title{spatialstore: Store commonly used spatial functions and lookups}
\description{
This package stores the commonly used spatial functions and lookups, along with function tests and documentation. These can be then used when creating new packages by copy and pasting the relevant R scripts and associated test scripts into the new package, as well as updating the pkgdown file
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gcfrench.github.io/spatialstore}
}

}
\author{
\strong{Maintainer}: Graham French \email{gcfrench@googlemail.com}

}
\keyword{internal}
